skincare-analysis/backend/requirements.txt
flask
flask-cors
ultralytics
torch
Pillow